pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "drumcut"
version = "0.1.0"
description = "CLI tool for automating drum practice video preprocessing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "John Vivian", email = "jvivian@users.noreply.github.com"}
]
keywords = ["drums", "video", "audio", "preprocessing", "gopro", "reaper"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "typer>=0.9.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyloudnorm>=0.1.0",
    "ffmpeg-python>=0.2.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
drumcut = "drumcut.cli:app"

[project.urls]
Homepage = "https://github.com/jvivian/drumcut"
Repository = "https://github.com/jvivian/drumcut"
Issues = "https://github.com/jvivian/drumcut/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["drumcut*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["drumcut"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=drumcut --cov-report=term-missing"
